chore: rename test case title and hooks based on comment
