+++ packages/account/src/Sections/Profile/PhoneVerification/__test__/confirm-phone-number.spec.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { StoreProvider, mockStore } from '@deriv/stores';
-import { useGetPhoneNumberOTP } from '@deriv/hooks';
+import { useRequestPhoneNumberOTP } from '@deriv/hooks';
 import ConfirmPhoneNumber from '../confirm-phone-number';
 import { WS } from '@deriv/shared';
 
 jest.mock('@deriv/hooks', () => ({
     ...jest.requireActual('@deriv/hooks'),
-    useGetPhoneNumberOTP: jest.fn(() => ({
+    useRequestPhoneNumberOTP: jest.fn(() => ({
         error_message: '',
         validatePhoneNumber: jest.fn(),
         requestOnWhatsApp: jest.fn(),
@@ -58,9 +58,9 @@ describe('ConfirmPhoneNumber', () => {
         expect(screen.getByRole('button', { name: 'Get code via WhatsApp' })).toBeInTheDocument();
     });
 
-    it('should render validatePhoneNumber function on user key pressed', () => {
+    it('should call validatePhoneNumber when the user presses a key', () => {
         const mock_validate_phone_number = jest.fn();
-        (useGetPhoneNumberOTP as jest.Mock).mockReturnValue({
+        (useRequestPhoneNumberOTP as jest.Mock).mockReturnValue({
             validatePhoneNumber: mock_validate_phone_number,
         });
         render(
@@ -76,8 +76,8 @@ describe('ConfirmPhoneNumber', () => {
         expect(mock_validate_phone_number).toHaveBeenCalled();
     });
 
-    it('should render display given error message', () => {
-        (useGetPhoneNumberOTP as jest.Mock).mockReturnValue({
+    it('should display given error message', () => {
+        (useRequestPhoneNumberOTP as jest.Mock).mockReturnValue({
             error_message: 'This is an error message',
         });
         render(
@@ -93,7 +93,7 @@ describe('ConfirmPhoneNumber', () => {
         (WS.authorized.setSettings as jest.Mock).mockReturnValue({
             error: { code: 'SomeErrorCode', message: 'SomeErrorMessage' },
         });
-        (useGetPhoneNumberOTP as jest.Mock).mockReturnValue({
+        (useRequestPhoneNumberOTP as jest.Mock).mockReturnValue({
             requestOnWhatsApp: jest.fn(),
             handleError: mock_handle_error,
         });
@@ -111,7 +111,7 @@ describe('ConfirmPhoneNumber', () => {
     it('should call requestOnWhatsApp when Whatsapp button is clicked', async () => {
         const mockWhatsappButtonClick = jest.fn();
 
-        (useGetPhoneNumberOTP as jest.Mock).mockReturnValue({
+        (useRequestPhoneNumberOTP as jest.Mock).mockReturnValue({
             requestOnWhatsApp: mockWhatsappButtonClick,
             handleError: jest.fn(),
         });
@@ -128,7 +128,7 @@ describe('ConfirmPhoneNumber', () => {
     it('should call requestOnSMS when SMS button is clicked', async () => {
         const mockSmsButtonClick = jest.fn();
 
-        (useGetPhoneNumberOTP as jest.Mock).mockReturnValue({
+        (useRequestPhoneNumberOTP as jest.Mock).mockReturnValue({
             requestOnSMS: mockSmsButtonClick,
             handleError: jest.fn(),
         });
