+++ packages/hooks/src/index.ts
@@ -23,7 +23,7 @@ export { default as useFiatAccountList } from './useFiatAccountList';
 export { default as useFileUploader } from './useFileUploader';
 export { default as useGetMFAccountStatus } from './useGetMFAccountStatus';
 export { default as useGetPasskeysList } from './useGetPasskeysList';
-export { default as useGetPhoneNumberOTP } from './useGetPhoneNumberOTP';
+export { default as useRequestPhoneNumberOTP } from './useRequestPhoneNumberOTP';
 export { default as useHasActiveRealAccount } from './useHasActiveRealAccount';
 export { default as useHasCryptoCurrency } from './useHasCryptoCurrency';
 export { default as useHasFiatCurrency } from './useHasFiatCurrency';
